+++ source/common/http/websocket/ws_handler_impl.cc
@@ -42,8 +42,7 @@ TcpProxy::ConfigSharedPtr Config(const envoy::api::v2::route::RouteAction& route
   return std::make_shared<TcpProxy::Config>(tcp_config, factory_context);
 }
 
-WsHandlerImpl::WsHandlerImpl(HeaderMap& request_headers,
-                             const RequestInfo::RequestInfo& request_info,
+WsHandlerImpl::WsHandlerImpl(HeaderMap& request_headers, RequestInfo::RequestInfo& request_info,
                              const Router::RouteEntry& route_entry,
                              WebSocketProxyCallbacks& callbacks,
                              Upstream::ClusterManager& cluster_manager,
@@ -143,6 +142,8 @@ void WsHandlerImpl::onConnectionSuccess() {
     ASSERT(queued_data_.length() == 0);
   }
 }
+
+RequestInfo::RequestInfo& WsHandlerImpl::getRequestInfo() { return request_info_; }
 
 } // namespace WebSocket
 } // namespace Http