+++ source/common/tcp_proxy/tcp_proxy.h
@@ -224,6 +224,8 @@ protected:
 
   virtual void onConnectionSuccess() {}
 
+  virtual RequestInfo::RequestInfo& getRequestInfo() { return request_info_; }
+
   void initialize(Network::ReadFilterCallbacks& callbacks, bool set_connection_stats);
   Network::FilterStatus initializeUpstreamConnection();
   void onConnectTimeout();