+++ test/integration/integration_test.cc
@@ -2,6 +2,9 @@
 
 #include <string>
 
+#include "envoy/config/accesslog/v2/file.pb.h"
+
+#include "common/filesystem/filesystem_impl.h"
 #include "common/http/header_map_impl.h"
 #include "common/protobuf/utility.h"
 
@@ -11,6 +14,8 @@
 #include "test/test_common/utility.h"
 
 #include "gtest/gtest.h"
+
+using testing::MatchesRegex;
 
 namespace Envoy {
 
@@ -377,6 +382,92 @@ TEST_P(IntegrationTest, WebSocketIdleTimeout) {
   test_server_->waitForCounterGe("tcp.my-stat-prefix.idle_timeout", 1);
   tcp_client->waitForDisconnect();
   fake_upstream_connection->waitForDisconnect();
+}
+
+TEST_P(IntegrationTest, WebSocketLogging) {
+  envoy::api::v2::route::RouteAction::WebSocketProxyConfig ws_config;
+  ws_config.mutable_idle_timeout()->set_nanos(
+      std::chrono::duration_cast<std::chrono::nanoseconds>(std::chrono::milliseconds(100)).count());
+  *ws_config.mutable_stat_prefix() = "my-stat-prefix";
+
+  config_helper_.setDefaultHostAndRoute("*", "/asd");
+  config_helper_.addConfigModifier(setRouteUsingWebsocket(&ws_config));
+
+  std::string expected_log_template = "bytes_sent={0} "
+                                      "bytes_received={1} "
+                                      "downstream_local_address={2} "
+                                      "downstream_remote_address={3} "
+                                      "upstream_local_address={4}";
+
+  std::string access_log_path = TestEnvironment::temporaryPath(fmt::format(
+      "websocket_access_log{}.txt", GetParam() == Network::Address::IpVersion::v4 ? "v4" : "v6"));
+  config_helper_.addConfigModifier([&](envoy::config::bootstrap::v2::Bootstrap& bootstrap) -> void {
+    auto* listener = bootstrap.mutable_static_resources()->mutable_listeners(0);
+    auto* filter_chain = listener->mutable_filter_chains(0);
+    auto* config_blob = filter_chain->mutable_filters(0)->mutable_config();
+
+    envoy::config::filter::network::http_connection_manager::v2::HttpConnectionManager
+        http_conn_manager_config;
+    MessageUtil::jsonConvert(*config_blob, http_conn_manager_config);
+
+    auto* access_log = http_conn_manager_config.add_access_log();
+    access_log->set_name("envoy.file_access_log");
+    envoy::config::accesslog::v2::FileAccessLog access_log_config;
+    access_log_config.set_path(access_log_path);
+    access_log_config.set_format(fmt::format(
+        expected_log_template, "%BYTES_SENT%", "%BYTES_RECEIVED%", "%DOWNSTREAM_LOCAL_ADDRESS%",
+        "%DOWNSTREAM_REMOTE_ADDRESS%", "%UPSTREAM_LOCAL_ADDRESS%"));
+
+    MessageUtil::jsonConvert(access_log_config, *access_log->mutable_config());
+    MessageUtil::jsonConvert(http_conn_manager_config, *config_blob);
+  });
+
+  initialize();
+
+  // WebSocket upgrade, send some data and disconnect downstream
+  IntegrationTcpClientPtr tcp_client;
+  FakeRawConnectionPtr fake_upstream_connection;
+  const std::string upgrade_req_str = "GET /websocket/test HTTP/1.1\r\nHost: host\r\nConnection: "
+                                      "keep-alive, Upgrade\r\nUpgrade: websocket\r\n\r\n";
+  const std::string upgrade_resp_str =
+      "HTTP/1.1 101 Switching Protocols\r\nConnection: Upgrade\r\nUpgrade: websocket\r\n\r\n";
+
+  tcp_client = makeTcpConnection(lookupPort("http"));
+  // Send websocket upgrade request
+  // The request path gets rewritten from /websocket/test to /websocket.
+  // The size of headers received by the destination is 228 bytes.
+  tcp_client->write(upgrade_req_str);
+  fake_upstream_connection = fake_upstreams_[0]->waitForRawConnection();
+  const std::string data = fake_upstream_connection->waitForData(228);
+  // Accept websocket upgrade request
+  fake_upstream_connection->write(upgrade_resp_str);
+  tcp_client->waitForData(upgrade_resp_str);
+  // Standard TCP proxy semantics post upgrade
+  tcp_client->write("hello");
+  // datalen = 228 + strlen(hello)
+  fake_upstream_connection->waitForData(233);
+  fake_upstream_connection->write("world");
+  tcp_client->waitForData(upgrade_resp_str + "world");
+
+  fake_upstream_connection->close();
+  fake_upstream_connection->waitForDisconnect();
+
+  tcp_client->waitForDisconnect();
+  tcp_client->close();
+
+  std::string log_result;
+  do {
+    log_result = Filesystem::fileReadToEnd(access_log_path);
+  } while (log_result.empty());
+
+  const std::string ip_port_regex = (GetParam() == Network::Address::IpVersion::v4)
+                                        ? R"EOF(127\.0\.0\.1:[0-9]+)EOF"
+                                        : R"EOF(\[::1\]:[0-9]+)EOF";
+
+  EXPECT_THAT(log_result, MatchesRegex(fmt::format(expected_log_template,
+                                                   82, // response length
+                                                   5,  // hello length
+                                                   ip_port_regex, ip_port_regex, ip_port_regex)));
 }
 
 TEST_P(IntegrationTest, TestBind) {